+++ lib/src/festival/solar_festival.dart
@@ -1,61 +1,36 @@
-import '../abstract_tyme.dart';
-import '../enums/festival_type.dart';
+import '../evt/event.dart';
 import '../solar/solar_day.dart';
+import 'abstract_festival.dart';
 
 /// 公历现代节日
 ///
 /// Author: 6tail
-class SolarFestival extends AbstractTyme {
+class SolarFestival extends AbstractFestival {
   static const List<String> names = ['元旦', '妇女节', '植树节', '劳动节', '青年节', '儿童节', '建党节', '建军节', '教师节', '国庆节'];
 
-  static String data = '@00001011950@01003081950@02003121979@03005011950@04005041950@05006011950@06007011941@07008011933@08009101985@09010011950';
+  static String data = '0VV__0Ux0Xc__0Ux0Xg__0_Q0ZV__0Ux0ZY__0Ux0aV__0Ux0bV__0Uo0cV__0Ug0de__0_V0eV__0Ux';
 
-  /// 类型
-  final FestivalType type;
-
-  /// 索引
-  final int index;
-
-  /// 公历日
-  final SolarDay day;
-
-  /// 名称
-  late final String name;
-
-  /// 起始年
-  final int startYear;
-
-  SolarFestival(this.type, this.day, this.startYear, String data) : index = int.parse(data.substring(1, 3), radix: 10) {
-    name = names[index];
-  }
+  SolarFestival(super.index, super.event, SolarDay super.day);
 
   static SolarFestival? fromIndex(int year, int index) {
     if (index < 0 || index >= names.length) {
       return null;
     }
-    RegExp pattern = RegExp('@${index.toString().padLeft(2, '0')}\\d{9}');
-    RegExpMatch? match = pattern.firstMatch(data);
-    if (match == null) {
-      return null;
-    }
-    String dt = match.group(0)!;
-    FestivalType type = FestivalType.fromCode(dt.codeUnitAt(3) - '0'.codeUnitAt(0))!;
-    if (type != FestivalType.DAY) {
-      return null;
-    }
-    int startYear = int.parse(dt.substring(8), radix: 10);
-    return year < startYear ? null : SolarFestival(type, SolarDay(year, int.parse(dt.substring(4, 6), radix: 10), int.parse(dt.substring(6, 8), radix: 10)), startYear, dt);
+    int start = index * 8;
+    Event e = Event(names[index], '@${data.substring(start, start + 8)}');
+    return year < e.getStartYear() ? null : SolarFestival(index, e, SolarDay(year, e.getValue(2), e.getValue(3)));
   }
 
   static SolarFestival? fromYmd(int year, int month, int day) {
-    RegExp pattern = RegExp('@\\d{2}0${month.toString().padLeft(2, '0')}${day.toString().padLeft(2, '0')}\\d{4}');
-    RegExpMatch? match = pattern.firstMatch(data);
-    if (match == null) {
-      return null;
+    SolarDay d = SolarDay(year, month, day);
+    for (int i = 0, j = names.length; i < j; i++) {
+      int start = i * 8;
+      Event e = Event(names[i], '@${data.substring(start, start + 8)}');
+      if (d.year >= e.getStartYear() && d.month == e.getValue(2) && d.day == e.getValue(3)) {
+        return SolarFestival(i, e, d);
+      }
     }
-    String dt = match.group(0)!;
-    int startYear = int.parse(dt.substring(8), radix: 10);
-    return year < startYear ? null : SolarFestival(FestivalType.DAY, SolarDay(year, month, day), startYear, dt);
+    return null;
   }
 
   @override
@@ -65,17 +40,8 @@ class SolarFestival extends AbstractTyme {
     return fromIndex((day.getYear() * size + i) ~/ size, indexOfSize(i, size))!;
   }
 
-  FestivalType getType() => type;
-
-  int getIndex() => index;
-
-  SolarDay getDay() => day;
-
-  int getStartYear() => startYear;
-
   @override
-  String getName() => name;
+  SolarDay getDay() => day as SolarDay;
 
-  @override
-  String toString() => '$day $name';
+  int getStartYear() => event.getStartYear();
 }