+++ lib/src/lunar/lunar_week.dart
@@ -6,8 +6,6 @@ import 'lunar_month.dart';
 ///
 /// Author: 6tail
 class LunarWeek extends WeekUnit {
-  static const List<String> names = ['第一周', '第二周', '第三周', '第四周', '第五周', '第六周'];
-
   LunarWeek(int year, int month, int index, int start): super(year, month, index, start) {
     validate(year, month, index, start);
   }
@@ -26,7 +24,7 @@ class LunarWeek extends WeekUnit {
   LunarMonth getLunarMonth() => LunarMonth(year, month);
 
   @override
-  String getName() => names[index];
+  String getName() => WeekUnit.names[index];
 
   @override
   String toString() => '${getLunarMonth()}${getName()}';