+++ lib/src/tyme.dart
@@ -4,6 +4,6 @@ import 'culture.dart';
 ///
 /// Author: 6tail
 abstract class Tyme implements Culture {
-  /// 推移[n]步
+  /// 推移[n]步，正数顺推，负数逆推
   Tyme? next(int n);
 }