+++ lib/src/unit/week_unit.dart
@@ -1,9 +1,12 @@
+import '../abstract_culture.dart';
 import 'month_unit.dart';
 
 /// 周
 ///
 /// Author: 6tail
 abstract class WeekUnit extends MonthUnit {
+  static const List<String> names = ['第一周', '第二周', '第三周', '第四周', '第五周', '第六周'];
+
   /// 索引，0-5
   final int index;
 
@@ -19,11 +22,7 @@ abstract class WeekUnit extends MonthUnit {
   int getStart() => start;
 
   static void validate(int index, int start) {
-    if (index < 0 || index > 5) {
-      throw ArgumentError('illegal week index: $index');
-    }
-    if (start < 0 || start > 6) {
-      throw ArgumentError('illegal week start: $start');
-    }
+    AbstractCulture.validateRange(index, 0, 5, 'week index');
+    AbstractCulture.validateRange(start, 0, 6, 'week start');
   }
 }