+++ CHANGELOG.md
@@ -31,3 +31,14 @@
 3. Added: Event.
 4. Optimized: Code and algorithms.
 5. Modified: Names of SolarFestival and LegalHoliday.
+
+## [1.4.4] - 2026-05-12
+1. Removed: FestivalType.
+2. Optimized: SolarFestival and LunarFestival are now implemented by Event.
+3. Optimized: SolarWeek and LunarWeek share the same name. 
+4. Optimized: RabByungMonth uses the name of LunarMonth. 
+5. Optimized: RabByungElement use separate names.
+
+## [1.5.0] - 2026-06-28
+1. Added：Hijri calendar。
+2. Optimized：Code and algorithms.